+++ Coco/src/Core/Rendering/RendererAPI.h
@@ -25,7 +25,7 @@ namespace Coco
 		virtual void Clear() = 0;
 		virtual void SetClearColor(const glm::vec4 color) = 0;
 
-		virtual void DrawIndexed(const std::shared_ptr<VertexArray>&vertexArray, uint32_t indexCount = 0) = 0;
+		virtual void DrawIndexed(const std::shared_ptr<VertexArray>&vertexArray, size_t indexCount = 0) = 0;
 
 		static void Create();
 		static void Destroy();
