+++ Coco/src/Platform/OpenGL/OpenGLRendererAPI.h
@@ -17,7 +17,7 @@ namespace Coco
 		virtual void Clear() override;
 		virtual void SetClearColor(const glm::vec4 color) override;
 
-		virtual void DrawIndexed(const Ref<VertexArray>& vertexArray, uint32_t indexCount) override;
+		virtual void DrawIndexed(const Ref<VertexArray>& vertexArray, size_t indexCount) override;
   };
 }
